print tags when listing available pipelines; split up files in pipeline/
